Fix build and warnings on Linux: clang 3.9, gcc 6

+++ core/context2D_data/DataProfile2DContextPlot.cpp
@@ -182,7 +182,7 @@ void DataProfile2DContextPlot::updatePlot()
     auto profilePoints = profileDataSet->GetPoints()->GetData();
     for (vtkIdType i = 0; i < numPoints; ++i)
     {
-        xAxis->SetValue(i, profilePoints->GetComponent(i, 0));
+        xAxis->SetValue(i, static_cast<float>(profilePoints->GetComponent(i, 0)));
     }
 
     table->AddColumn(xAxis);
